Makes plugin's install function use options when passed by the user.
Instead of discarding the available second argument to Vue.use(), use it for
providing default values to the component instance, when needed.
Both the props `options` and `emit` can be configured passing an object like this:
Vue.use(VuePlyr, {
  plyr: {
    ratio: '4:5',
    loop: { active: true }
  },
  emit: [ 'progress', 'ended' ]
})

The values are set as prop default so they can easily be overwritten when
instantiating the component.
I use Plyr with localized messages and I found super convenient to use this
method to only import and set the messages once when installing the plugin.